migrate current css to scss, and moved _shared to scss partials folder
