+++ client/src/components/Header/Header.scss
@@ -1,3 +1,5 @@
+@import '../../scss/partials/button';
+
 .header {
   padding: 0 60px;
   margin-top: 60px;
@@ -6,7 +8,7 @@
 }
 
 .button {
-  composes: button from '../_shared/button.css';
+  @extend %button;
   padding: 10px;
   margin-left: 10px;
 }
