+++ client/src/components/Signup/Signup.js
@@ -4,7 +4,7 @@ import { Field, reduxForm } from 'redux-form'
 import { push } from 'react-router-redux'
 import PropTypes from 'prop-types'
 
-import styles from '../Login/Login.css'
+import styles from '../Login/Login.scss'
 import { signup } from '../../actions'
 
 class Signup extends Component {
